+++ src/main/java/com/example/heroku/model/joinwith/UserPackageDetailJoinWithUserPackage.java
@@ -60,7 +60,7 @@ public class UserPackageDetailJoinWithUserPackage {
     private String child_device_id;
     private String child_product_second_id;
     private String child_product_unit_second_id;
-    private int child_number_unit;
+    private float child_number_unit;
     private float child_price;
     private float child_buy_price;
     private float child_discount_amount;
